Run warning-free in HomeBridge
I was able to add the accessory to homebridge and then interact with it
as expected in the Home app. This generated a couple of warnings around
the range of the static temperature values I was setting (have to use
celcius for all values) and return values, which I cleaned up.