Implement temperature units in the API
This allows reading and setting the thermostat temperature display units. I've verified that these methods read and write the data appropriately. Changing the unit via the API does update the accessory settings visible in the home app, although the tile remains in Fahrenheit. I think that part might have to do with the locale of the device though.

Closes #3